+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * (breaking) remove `Goal`, `View` as top level statements
 * (breaking) IEP 16 (intensional objects).  Use new `$` built-in function to evaluate the symbol to be applied.
 * IEP 03: allow enumeration of constructed types in theory/structure block
+* IEP 03 : partial support for binary quantification, with unary predicates
 * fix #70: Double unary symbol should be parsed
 * Fix #72: Empty predicate in structure should be allowed